+++ TreasureData/TreasureData.h
@@ -70,7 +70,7 @@ typedef void (^ErrorHandler)(NSString* _Nonnull errorCode, NSString* _Nullable e
  *    AWS East  https://cdp.in.treasuredata.com
  *    AWS Tokyo https://cdp-tokyo.in.treasuredata.com
  *    AWS EU    https://cdp-eu01.in.treasuredata.com
- *    IDCF      https://cdp-idcf.in.treasuredata.com
+ *    AWS Asia Pacific (Seoul)  https://cdp-ap02.in.treasuredata.com
  */
 @property(nonatomic, strong) NSString * _Nullable cdpEndpoint;
 
@@ -78,6 +78,11 @@ typedef void (^ErrorHandler)(NSString* _Nonnull errorCode, NSString* _Nullable e
 
 /**
  * Assign the target API endpoint, default is "https://in.treasuredata.com".
+ * Possible values:
+ *    AWS East  https://in.treasuredata.com
+ *    AWS Tokyo https://tokyo.in.treasuredata.com
+ *    AWS EU    https://eu01.in.treasuredata.com
+ *    AWS Asia Pacific (Seoul)  https://ap02.in.treasuredata.com
  * This have to be call before `initializeWithApiKey(apiKey:)`, otherwise it won't have effect.
  * @param apiEndpoint for the in effect endpoint (`+[TreasureData initializeApiEndpoint:]`).
  */
@@ -182,6 +187,11 @@ typedef void (^ErrorHandler)(NSString* _Nonnull errorCode, NSString* _Nullable e
 - (void)uploadEvents;
 
 #pragma mark - Events' metadata
+
+/**
+ * Get UUID generated from TreasureData. The value will be set to `td_uuid` column for every events if `enableAutoAppendUniqId` is called.
+ */
+- (NSString *_Nonnull)getUUID;
 
 /**
  * Automaticaly append `td_uuid` column for every events. The value is randomly generated and persisted, it is shared across app launches and events. Basically, it is used to prepresent for a unique app installation instance.