+++ TreasureDataTests/IntegrationTests.swift
@@ -14,6 +14,7 @@ class IntegrationTests: XCTestCase {
     static var api: TDAPI!
     static var apiKey: String!
     static var collectorEndpoint: String!
+    static var cdpEndpoint: String!
 
     static let TargetDatabase = "ios_it"
     static var sessionPrefix = uuid()
@@ -33,6 +34,9 @@ class IntegrationTests: XCTestCase {
         guard let collectorEndpoint = ProcessInfo.processInfo.environment["COLLECTOR_ENDPOINT"] else {
             fatalError("Missing env COLLECTOR_ENDPOINT")
         }
+        guard let cdpEndpoint = ProcessInfo.processInfo.environment["CDP_ENDPOINT"] else {
+            fatalError("Missing env CDP_ENDPOINT")
+        }
         // Syntax: AUDIENCE_TOKENS=token1,token2
         guard let audienceTokensString = ProcessInfo.processInfo.environment["AUDIENCE_TOKENS"] else {
             fatalError("Missing env AUDIENCE_TOKENS")
@@ -44,6 +48,7 @@ class IntegrationTests: XCTestCase {
         
         IntegrationTests.apiKey = apiKey
         IntegrationTests.collectorEndpoint = collectorEndpoint
+        IntegrationTests.cdpEndpoint = cdpEndpoint
         IntegrationTests.audienceTokens = audienceTokensString.split(separator: ",").map { String($0) }
         let keyValueStrings = userSegmentKeysString.split(separator: ",").map { String($0) }
         keyValueStrings.forEach { keyValueString in
@@ -65,6 +70,7 @@ class IntegrationTests: XCTestCase {
         TreasureData.initializeApiEndpoint(IntegrationTests.collectorEndpoint)
         sdkClient = TreasureData(apiKey: IntegrationTests.apiKey)
         sdkClient.defaultDatabase = IntegrationTests.TargetDatabase
+        sdkClient.cdpEndpoint = IntegrationTests.cdpEndpoint
         tempTables = []
     }
 
@@ -94,18 +100,30 @@ class IntegrationTests: XCTestCase {
         }
         
     }
-
+    
     func testLifeCycleAppOpenedEvent() {
         sdkClient.enableAppLifecycleEvent()
         sdkClient.defaultTable = newTempTable()
         NotificationCenter.default.post(name: Notification.Name("UIApplicationDidFinishLaunchingNotification"), object: nil)
         sdkClient.uploadEvents()
         let result = try! IntegrationTests.api.stubbornQuery(
-            "select td_ios_event, td_app_ver, td_app_ver_num from \(sdkClient.defaultTable!) limit 1",
+            "select td_ios_event, td_app_ver, td_app_ver_num from \(sdkClient.defaultTable!) limit 2",
             database: IntegrationTests.TargetDatabase)
-        XCTAssertEqual(result[0][0] as! String, "TD_IOS_APP_OPEN")
-        XCTAssertNotNil(result[0][1])
-        XCTAssertNotNil(result[0][2])
+        
+        XCTAssert(result.count == 1 || result.count == 2, "Invalid number of events")
+        if result.count == 1 {
+            XCTAssertEqual(result[0][0] as! String, "TD_IOS_APP_OPEN")
+            XCTAssertNotNil(result[0][1])
+            XCTAssertNotNil(result[0][2])
+        } else if result.count == 2 {
+            let event0 = result[0][0] as! String
+            let event1 = result[1][0] as! String
+            XCTAssert(event0 == "TD_IOS_APP_OPEN" || event1 == "TD_IOS_APP_OPEN", "No TD_IOS_APP_OPEN event found")
+            XCTAssertNotNil(result[0][1])
+            XCTAssertNotNil(result[0][2])
+            XCTAssertNotNil(result[1][1])
+            XCTAssertNotNil(result[1][2])
+        }
     }
 
     func testSession() {